% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{apply_data_corrections}
\alias{apply_data_corrections}
\title{Apply Data Corrections}
\usage{
apply_data_corrections(con, verbose = TRUE)
}
\arguments{
\item{con}{DuckDB connection}

\item{verbose}{Print correction messages (default: TRUE)}
}
\value{
Invisibly returns the connection
}
\description{
Applies known data corrections to the database. This function should be
updated as new corrections are identified by data managers.
}
\examples{
\dontrun{
con <- get_duckdb_con()
apply_data_corrections(con)
}
}
\concept{wrangle}
