% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitemap.R
\name{assert_sitemap}
\alias{assert_sitemap}
\title{Stop when \code{\link[=check_sitemap]{check_sitemap()}} found an error}
\usage{
assert_sitemap(d, quiet = FALSE)
}
\arguments{
\item{d}{the tibble from \code{\link[=check_sitemap]{check_sitemap()}}}

\item{quiet}{suppress the summary line}
}
\value{
\code{d}, invisibly.
}
\description{
Stop when \code{\link[=check_sitemap]{check_sitemap()}} found an error
}
\concept{catalog}
