% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac.R
\name{assert_stac}
\alias{assert_stac}
\title{Stop when \code{\link[=check_stac]{check_stac()}} found an error}
\usage{
assert_stac(d, quiet = FALSE)
}
\arguments{
\item{d}{the tibble from \code{\link[=check_stac]{check_stac()}}}

\item{quiet}{suppress the summary line}
}
\value{
\code{d}, invisibly.
}
\description{
Stop when \code{\link[=check_stac]{check_stac()}} found an error
}
\concept{catalog}
