% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\docType{data}
\name{CC_RELEASE_PREFIX}
\alias{CC_RELEASE_PREFIX}
\alias{CC_TABLES_PREFIX}
\title{Release layout prefixes (bucket-relative)}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
CC_RELEASE_PREFIX

CC_TABLES_PREFIX
}
\description{
Release layout prefixes (bucket-relative)
}
\concept{release}
\keyword{datasets}
