% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint.R
\name{changed_inputs}
\alias{changed_inputs}
\title{Which inputs changed since a recorded fingerprint}
\usage{
changed_inputs(fp, prior)
}
\arguments{
\item{fp}{output of \code{\link[=input_fingerprint]{input_fingerprint()}}}

\item{prior}{output of \code{\link[=read_input_fingerprint]{read_input_fingerprint()}}; \code{NULL} means everything is
new}
}
\value{
character vector of input names that were added, removed or changed
}
\description{
Which inputs changed since a recorded fingerprint
}
\concept{ingest}
