% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{close_duckdb}
\alias{close_duckdb}
\title{Disconnect from DuckDB and shutdown}
\usage{
close_duckdb(con, checkpoint = NULL)
}
\arguments{
\item{con}{DuckDB connection}

\item{checkpoint}{Run checkpoint before closing (default: TRUE for writeable connections)}
}
\value{
NULL (invisibly)
}
\description{
Properly closes a DuckDB connection and shuts down the database.
}
\examples{
\dontrun{
close_duckdb(con)
}
}
\concept{duckdb}
