% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{determine_field_types}
\alias{determine_field_types}
\title{Determine Field Types for Database}
\usage{
determine_field_types(d, tbl, fld)
}
\arguments{
\item{d}{Data frame with raw data}

\item{tbl}{Table name}

\item{fld}{Field name}
}
\value{
Character vector of PostgreSQL data types
}
\description{
Determines appropriate PostgreSQL data types for fields based on data values.
}
\examples{
\dontrun{
determine_field_types(raw_data, "species", "species_name")
}
}
