% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{download_archive}
\alias{download_archive}
\title{Download archive to local directory}
\usage{
download_archive(
  archive_timestamp,
  provider,
  dataset,
  local_dir = NULL,
  gcs_bucket = "calcofi-files-public",
  archive_prefix = "archive"
)
}
\arguments{
\item{archive_timestamp}{Archive timestamp}

\item{provider}{Data provider}

\item{dataset}{Dataset name}

\item{local_dir}{Local directory to download to (default: temp directory)}

\item{gcs_bucket}{GCS bucket name}

\item{archive_prefix}{Archive folder prefix}
}
\value{
Path to local directory containing downloaded files
}
\description{
Downloads CSV files from a GCS archive to a local directory.
}
\examples{
\dontrun{
local_dir <- download_archive(
  archive_timestamp = "2026-02-02_121557",
  provider = "swfsc.noaa.gov",
  dataset = "calcofi-db")
}
}
\concept{archive}
