% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ship.R
\name{fetch_ship_ices}
\alias{fetch_ship_ices}
\title{Fetch Ship Codes from ICES Reference Codes API}
\usage{
fetch_ship_ices(
  ices_api = "https://vocab.ices.dk/services/api",
  ices_ship_code_type = "7f9a91e1-fb57-464a-8eb0-697e4b0235b5"
)
}
\arguments{
\item{ices_api}{character; ICES API base URL}

\item{ices_ship_code_type}{character; GUID for ship code type}
}
\value{
tibble with columns: ship_nodc, ship_name, remarks, src
}
\description{
Retrieves the full list of ship/platform codes from the ICES Vocabularies
API. The default code type GUID corresponds to "SHIPC" (ship codes).
}
\examples{
\dontrun{
d_ices <- fetch_ship_ices()
}
}
\concept{ship}
