% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{license_statuses}
\alias{license_statuses}
\title{The allowed \code{status} values of \code{metadata/license.csv}}
\usage{
license_statuses()
}
\value{
Character vector.
}
\description{
The allowed \code{status} values of \code{metadata/license.csv}
}
\concept{registry}
