% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{list_gcs_files}
\alias{list_gcs_files}
\title{List files in a GCS bucket/prefix}
\usage{
list_gcs_files(bucket, prefix = NULL, recursive = TRUE)
}
\arguments{
\item{bucket}{GCS bucket name}

\item{prefix}{Path prefix to filter results}

\item{recursive}{Whether to list recursively (default: TRUE)}
}
\value{
Data frame with columns: name, size, updated, md5Hash
}
\description{
Lists objects in a GCS bucket with optional prefix filter.
}
\examples{
\dontrun{
# list all files in current/
files <- list_gcs_files("calcofi-files", prefix = "current/")

# list bottle-database files
files <- list_gcs_files(
  "calcofi-files",
  prefix = "current/calcofi.org/bottle-database/")
}
}
\concept{cloud}
