% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{list_gcs_versions}
\alias{list_gcs_versions}
\title{List versions of a file in GCS archive}
\usage{
list_gcs_versions(path, bucket = "calcofi-files-public")
}
\arguments{
\item{path}{Relative path to the file (e.g., "calcofi.org/bottle-database/bottle.csv")}

\item{bucket}{GCS bucket name (default: "calcofi-files")}
}
\value{
Data frame with columns: version_date, gcs_path, size, updated
}
\description{
Finds all archived versions of a file in the calcofi-files bucket.
}
\examples{
\dontrun{
versions <- list_gcs_versions("calcofi.org/bottle-database/bottle.csv")
}
}
\concept{cloud}
