% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ns_key}
\alias{ns_key}
\title{Namespaced \code{sample_key} expression: \code{dataset_key:sample_type:id}}
\usage{
ns_key(dataset_key, sample_type, id_sql)
}
\arguments{
\item{dataset_key, sample_type, id_sql}{components}
}
\value{
a SQL expression string
}
\description{
Namespaced \code{sample_key} expression: \code{dataset_key:sample_type:id}
}
\concept{model}
