% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_csv_metadata}
\alias{read_csv_metadata}
\title{Read CSV Files and Extract Metadata}
\usage{
read_csv_metadata(dir_csv, dir_ingest, create_dirs = TRUE)
}
\arguments{
\item{dir_csv}{Directory containing CSV files}

\item{dir_ingest}{Directory to store metadata files}

\item{create_dirs}{Whether to create directories if they don't exist}
}
\value{
A list with two data frames: tables and fields metadata
}
\description{
Reads all CSV files in a directory, extracts metadata about tables and fields,
and writes this metadata to files for further processing.
}
\examples{
\dontrun{
csv_metadata <- read_csv_metadata(
  dir_csv = "/path/to/data/provider/dataset",
  dir_ingest = "/path/to/ingest/provider/dataset")
}
}
\concept{read}
