% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry_measurement_face.R
\name{read_measurement_chem}
\alias{read_measurement_chem}
\title{Read \code{metadata/measurement_chem.csv}}
\usage{
read_measurement_chem(path, validate = TRUE)
}
\arguments{
\item{path}{path to \code{metadata/measurement_chem.csv}}

\item{validate}{error on sentinel strings (default TRUE)}
}
\value{
tibble
}
\description{
Read \code{metadata/measurement_chem.csv}
}
\concept{registry}
