% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd_upload.R
\name{read_sbe_asc}
\alias{read_sbe_asc}
\title{Read a Sea-Bird \code{.asc} ASCII export}
\usage{
read_sbe_asc(path)
}
\arguments{
\item{path}{file path}
}
\value{
a tibble with a \code{sbe_header} attribute
}
\description{
Read a Sea-Bird \code{.asc} ASCII export
}
\concept{upload}
