% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{RELEASE_REQUIRED_OBJECTS}
\alias{RELEASE_REQUIRED_OBJECTS}
\title{Objects a frozen release must contain before it can be promoted}
\usage{
RELEASE_REQUIRED_OBJECTS
}
\description{
Objects a frozen release must contain before it can be promoted
}
\keyword{internal}
