% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{show_googledrive_files}
\alias{show_googledrive_files}
\title{Show Google Drive files}
\usage{
show_googledrive_files(d)
}
\arguments{
\item{d}{data object output from \code{read_csv_files()}}
}
\value{
Data frame with Google Drive files
}
\description{
Show Google Drive files
}
\concept{viz}
