% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{show_source_files}
\alias{show_source_files}
\title{Show source files}
\usage{
show_source_files(d)
}
\arguments{
\item{d}{data object output from \code{read_csv_files()}}
}
\value{
DT datatable showing source files with paths and row counts
}
\description{
Show source CSV files with local and GCS paths for provenance
}
\concept{viz}
