% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{show_tables_redefine}
\alias{show_tables_redefine}
\title{Show tables to redefine}
\usage{
show_tables_redefine(d)
}
\arguments{
\item{d}{data object output from \code{read_csv_files()}}
}
\value{
Data frame with tables to redefine
}
\description{
Show tables to redefine
}
\concept{viz}
