% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitemap.R
\name{write_sitemap_xml}
\alias{write_sitemap_xml}
\title{Write a sitemap XML file (sitemaps.org 0.9)}
\usage{
write_sitemap_xml(d, path)
}
\arguments{
\item{d}{the tibble from \code{\link[=build_datasets_sitemap]{build_datasets_sitemap()}}}

\item{path}{the file to write}
}
\value{
\code{path}, invisibly.
}
\description{
Write a sitemap XML file (sitemaps.org 0.9)
}
\concept{catalog}
