% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcofi4db-package.R
\docType{package}
\name{calcofi4db-package}
\alias{calcofi4db}
\alias{calcofi4db-package}
\title{calcofi4db: CalCOFI Database Tools}
\description{
Tools for ingesting datasets into the CalCOFI database with functions to read source CSV files, detect changes in tables and fields, and update the database with proper metadata.
}
\author{
\strong{Maintainer}: Ben Best \email{ben@ecoquants.com} (\href{https://orcid.org/0000-0002-2686-0784}{ORCID})

}
\keyword{internal}
