% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{CC_RELEASE_PREFIX}
\alias{CC_RELEASE_PREFIX}
\alias{CC_TABLES_PREFIX}
\title{Release layout prefixes (bucket-relative)}
\usage{
CC_RELEASE_PREFIX

CC_TABLES_PREFIX
}
\description{
Release layout prefixes (bucket-relative)
}
\concept{release}
