% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_datasets.R
\name{classify_portal}
\alias{classify_portal}
\title{Which portal family a URL belongs to}
\usage{
classify_portal(url)
}
\arguments{
\item{url}{character vector}
}
\value{
character vector of the same length, values from \code{\link[=distribution_portals]{distribution_portals()}}.
}
\description{
Host-based: \code{edi} (edirepository.org, pasta.lternet.edu), \code{ncei},
\code{erddap-noaa} (coastwatch / oceanview / upwell \code{pfeg.noaa.gov} ERDDAPs),
\code{erddap-calcofi}, \code{datazoo} and the other oceaninformatics.ucsd.edu portals
(ZooDB, ZooScan), \code{ucsd-library}, \code{obis}, \code{ipt}, \code{caloos}, \code{zenodo}, \code{ncbi},
\code{calcofi.org}, \code{gcs} (storage.googleapis.com / storage.calcofi.io), else
\code{other}. \code{NA} for an empty input.
}
\concept{catalog}
