% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_db_con}
\alias{get_db_con}
\title{Get a database connection to the CalCOFI PostgreSQL database}
\usage{
get_db_con(
  schemas = "public",
  host = NULL,
  port = 5432,
  dbname = "gis",
  user = "admin",
  password_file = NULL
)
}
\arguments{
\item{schemas}{Character vector of schema search paths (default: "public")}

\item{host}{Database host (default: automatically detected based on system)}

\item{port}{Database port (default: 5432)}

\item{dbname}{Database name (default: "gis")}

\item{user}{Database user (default: "admin")}

\item{password_file}{Path to file containing database password (default: system-dependent location)}
}
\value{
A database connection object
}
\description{
Get a database connection to the CalCOFI PostgreSQL database
}
\concept{utils}
