% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_csv_files}
\alias{read_csv_files}
\title{Read CSV Files and Their Metadata}
\usage{
read_csv_files(
  provider,
  dataset,
  dir_data = "~/My Drive/projects/calcofi/data",
  url_gdata =
    "https://drive.google.com/drive/u/0/folders/1xxdWa4mWkmfkJUQsHxERTp9eBBXBMbV7",
  use_gdrive = TRUE,
  email = "ben@ecoquants.com"
)
}
\arguments{
\item{provider}{Data provider (e.g., "swfsc.noaa.gov")}

\item{dataset}{Dataset name (e.g., "calcofi-db")}

\item{dir_data}{directory path of CalCOFI base data folder available locally,
with CSVs under {provider}/{dataset} directory. Default: "~/My
Drive/projects/calcofi/data"}

\item{url_gdata}{URL of CalCOFI base data folder in Google Drive (with CSVs
under {provider}/{dataset} directory) with metadata information on CSVs.
Default: \href{https://drive.google.com/drive/u/0/folders/1xxdWa4mWkmfkJUQsHxERTp9eBBXBMbV7}{data - Google Drive}}

\item{use_gdrive}{Whether to query Google Drive for metadata. Default: TRUEs}

\item{email}{Google Drive authentication email (if use_gdrive=TRUE). Default:
"ben@ecoquants.com"}
}
\value{
A list containing table data and metadata
}
\description{
Reads CSV files from a directory and prepares them for ingestion into a
database.
}
\examples{
\dontrun{
dataset_info <- read_csv_files(
  provider = "swfsc.noaa.gov",
  dataset  = "calcofi-db")
}
}
\concept{read}
